# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.1",
    "title": "Dr.One API",
    "description": "Dr.One server documentation"
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://dr-one-marketplace.onrender.com" }
  ],
...
  "paths": {
...
    "/goods/add": {
      "post": {
        "tags": ["Good"],
        "summary": "Add a new good to the store",
        "description": "Add a new good to the store",
        "operationId": "addGood",
        "requestBody": {
          "description": "Create a new good in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfull operations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or not filled all the fields"
          }
        }
      }
    },
    "/goods/delete/{goodId}": {
      "delete": {
        "tags": ["Good"],
        "summary": "Deletes a good from store",
        "description": "Delete a good from store",
        "operationId": "delById",
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goodId",
            "in": "path",
            "description": "Good Id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "404": {
            "description": "Invalid good id"
          }
        }
      }
    },
    "/goods/edit/{goodId}": {
      "put": {
        "tags": ["Good"],
        "summary": "Edits an existing good by id",
        "description": "Edit an existing good by id",
        "operationId": "editGoodById",
        "parameters": [
          {
            "name": "goodId",
            "in": "path",
            "description": "Good id to edit",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Edit an existing good in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully edited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or not filled all the fields"
          },
          "404": {
            "description": "Invalid good id"
          }
        }
      }
    },
    "/goods/update/{goodId}": {
      "patch": {
        "tags": ["Good"],
        "summary": "Updates a price and quantity of good",
        "description": "Update a price and quantity of good",
        "operationId": "updateGoodById",
        "parameters": [
          {
            "name": "goodId",
            "in": "path",
            "description": "Good id to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#components/schemas/Good"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/Good"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or not filled all the fields"
          },
          "404": {
            "description": "Invalid good id"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic admin:123456"
      },
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}